Рефакторит компонент
